Revert "net: usb: raw-ip: increment number of rx buffers"
This reverts commit 6eb3f9cd2f8923ddc614c6e507cd9eb6c44313cc.

+++ drivers/net/usb/raw_ip_net.c
@@ -43,7 +43,6 @@
 
 /* maximum interface number supported */
 #define MAX_INTFS	5
-#define RAWIP_RX_BUFS	16
 
 MODULE_LICENSE("GPL");
 
@@ -93,8 +92,8 @@ struct baseband_usb {
 		struct work_struct tx_work;
 	} usb;
 	/* re-usable rx urb */
-	struct urb *urb_r[RAWIP_RX_BUFS];
-	void *buff[RAWIP_RX_BUFS];
+	struct urb *urb_r;
+	void *buff;
 	/* suspend count */
 	int susp_count;
 };
@@ -153,7 +152,7 @@ static int baseband_usb_driver_probe(struct usb_interface *intf,
 
 static void baseband_usb_driver_disconnect(struct usb_interface *intf)
 {
-	int i, j;
+	int i;
 	struct urb *urb;
 
 	pr_debug("%s intf %p\n", __func__, intf);
@@ -198,10 +197,11 @@ static void baseband_usb_driver_disconnect(struct usb_interface *intf)
 				= (struct urb *) 0;
 		}
 		/* kill usb rx */
-		for (j = 0; j < RAWIP_RX_BUFS; j++)
-			usb_kill_urb(baseband_usb_net[i]->urb_r[j]);
-		baseband_usb_net[i]->usb.rx_urb	= (struct urb *) 0;
-
+		if (baseband_usb_net[i]->usb.rx_urb) {
+			usb_kill_urb(baseband_usb_net[i]->usb.rx_urb);
+			baseband_usb_net[i]->usb.rx_urb
+				= (struct urb *) 0;
+		}
 		/* mark interface as disconnected */
 		baseband_usb_net[i]->usb.interface
 			= (struct usb_interface *) 0;
@@ -215,7 +215,7 @@ static void baseband_usb_driver_disconnect(struct usb_interface *intf)
 static int baseband_usb_driver_suspend(struct usb_interface *intf,
 	pm_message_t message)
 {
-	int i, susp_count, j;
+	int i, susp_count;
 
 	pr_debug("%s intf %p\n", __func__, intf);
 
@@ -253,10 +253,13 @@ static int baseband_usb_driver_suspend(struct usb_interface *intf,
 			continue;
 		}
 		/* kill usb rx */
+		if (!baseband_usb_net[i]->usb.rx_urb) {
+			pr_debug("rx_usb already killed\n");
+			up(&baseband_usb_net[i]->sem);
+			continue;
+		}
 		pr_debug("%s: kill rx_urb {\n",__func__);
-		for (j = 0; j < RAWIP_RX_BUFS; j++)
-			usb_kill_urb(baseband_usb_net[i]->urb_r[j]);
-
+		usb_kill_urb(baseband_usb_net[i]->usb.rx_urb);
 		pr_debug("%s: kill rx_urb }\n",__func__);
 		baseband_usb_net[i]->usb.rx_urb = (struct urb *) 0;
 		/* cancel tx urb work (will restart after resume) */
@@ -315,6 +318,11 @@ static int baseband_usb_driver_resume(struct usb_interface *intf)
 			continue;
 		}
 		/* start usb rx */
+		if (baseband_usb_net[i]->usb.rx_urb) {
+			pr_debug("rx_usb already exists\n");
+			up(&baseband_usb_net[i]->sem);
+			continue;
+		}
 		err = usb_net_raw_ip_rx_urb_submit(baseband_usb_net[i]);
 		if (err < 0) {
 			pr_err("submit rx failed - err %d\n", err);
@@ -579,7 +587,7 @@ static int usb_net_raw_ip_rx_urb_submit(struct baseband_usb *usb)
 {
 	struct urb *urb;
 	void *buf;
-	int err, j;
+	int err;
 
 	pr_debug("usb_net_raw_ip_rx_urb_submit { usb %p\n", usb);
 
@@ -592,30 +600,32 @@ static int usb_net_raw_ip_rx_urb_submit(struct baseband_usb *usb)
 		pr_err("usb interface disconnected - not submitting rx urb\n");
 		return -EINVAL;
 	}
+	if (usb->usb.rx_urb) {
+		pr_err("previous urb still active\n");
+		return -EBUSY;
+	}
+	if (!usb->urb_r || !usb->buff) {
+		pr_err("no reusable rx urb found\n");
+		return -ENOMEM;
+	}
 
-	for (j = 0; j < RAWIP_RX_BUFS; j++) {
-		urb = usb->urb_r[j];
-		buf = usb->buff[j];
-		if (!urb || !buf) {
-			pr_err("no reusable rx urb found\n");
-			return -ENOMEM;
-		}
+	/* reuse rx urb */
+	urb = usb->urb_r;
+	buf = usb->buff;
+	usb_fill_bulk_urb(urb, usb->usb.device, usb->usb.pipe.bulk.in,
+		buf, USB_NET_BUFSIZ,
+		usb_net_raw_ip_rx_urb_comp,
+		usb);
+	urb->transfer_flags = 0;
 
-		usb_fill_bulk_urb(urb, usb->usb.device, usb->usb.pipe.bulk.in,
-				buf, USB_NET_BUFSIZ,
-				usb_net_raw_ip_rx_urb_comp,
-				usb);
-		urb->transfer_flags = 0;
-
-		/* submit rx urb */
-		usb_mark_last_busy(usb->usb.device);
-		usb->usb.rx_urb = urb;
-		err = usb_submit_urb(urb, GFP_ATOMIC);
-		if (err < 0) {
-			pr_err("usb_submit_urb() failed - err %d\n", err);
-			usb->usb.rx_urb = (struct urb *) 0;
-			return err;
-		}
+	/* submit rx urb */
+	usb_mark_last_busy(usb->usb.device);
+	usb->usb.rx_urb = urb;
+	err = usb_submit_urb(urb, GFP_ATOMIC);
+	if (err < 0) {
+		pr_err("usb_submit_urb() failed - err %d\n", err);
+		usb->usb.rx_urb = (struct urb *) 0;
+		return err;
 	}
 
 	pr_debug("usb_net_raw_ip_rx_urb_submit }\n");
@@ -625,7 +635,7 @@ static int usb_net_raw_ip_rx_urb_submit(struct baseband_usb *usb)
 static void usb_net_raw_ip_rx_urb_comp(struct urb *urb)
 {
 	struct baseband_usb *usb;
-	int i, err;
+	int i;
 	struct sk_buff *skb;
 	unsigned char *dst;
 	unsigned char ethernet_header[14] = {
@@ -641,13 +651,13 @@ static void usb_net_raw_ip_rx_urb_comp(struct urb *urb)
 		NET_IP_ETHERTYPE,
 	};
 
+	pr_debug("usb_net_raw_ip_rx_urb_comp { urb %p\n", urb);
+
 	/* check input */
 	if (!urb) {
 		pr_err("no urb\n");
 		return;
 	}
-	pr_debug("usb_net_raw_ip_rx_urb_comp { urb %p\n", urb);
-
 	usb = (struct baseband_usb *)urb->context;
 	i = usb->baseband_index;
 	switch (urb->status) {
@@ -662,10 +672,6 @@ static void usb_net_raw_ip_rx_urb_comp(struct urb *urb)
 		pr_info("%s: rx urb %p - link shutdown %d EPROTO\n",
 			__func__, urb, urb->status);
 		goto err_exit;
-	case -ENOENT:
-		pr_debug("%s: rx urb %p status %d\n",
-				__func__, urb, urb->status);
-		break;
 	default:
 		pr_info("%s: rx urb %p - status %d\n",
 			__func__, urb, urb->status);
@@ -737,20 +743,15 @@ static void usb_net_raw_ip_rx_urb_comp(struct urb *urb)
 		}
 	}
 
+	/* mark rx urb complete */
+	usb->usb.rx_urb = (struct urb *) 0;
+
 	/* do not submit urb if interface is suspending */
 	if (urb->status == -ENOENT)
 		return;
 
-	/* submit this rx urb again */
-	usb_mark_last_busy(usb->usb.device);
-	usb->usb.rx_urb = urb;
-	err = usb_submit_urb(urb, GFP_ATOMIC);
-	if (err < 0) {
-		pr_err("usb_submit_urb() failed - err %d\n", err);
-		usb->usb.rx_urb = NULL;
-		return;
-	}
-
+	/* submit next rx urb */
+	usb_net_raw_ip_rx_urb_submit(usb);
 	return;
 
 err_exit:
@@ -763,7 +764,6 @@ err_exit:
 
 static int usb_net_raw_ip_setup_rx_urb( struct baseband_usb *usb)
 {
-	int i;
 	pr_debug("usb_net_raw_ip_setup_rx_urb {\n");
 
 	/* check input */
@@ -771,21 +771,23 @@ static int usb_net_raw_ip_setup_rx_urb( struct baseband_usb *usb)
 		pr_err("%s: !usb\n", __func__);
 		return -EINVAL;
 	}
+	if (usb->urb_r) {
+		pr_err("%s: reusable rx urb already allocated\n", __func__);
+		return -EINVAL;
+	}
 
-	for (i = 0; i < RAWIP_RX_BUFS; i++) {
-		/* allocate reusable rx urbs */
-		usb->urb_r[i] = usb_alloc_urb(0, GFP_KERNEL);
-		if (!usb->urb_r[i]) {
-			pr_err("usb_alloc_urb() failed\n");
-			return -ENOMEM;
-		}
-		usb->buff[i] = kzalloc(USB_NET_BUFSIZ, GFP_KERNEL);
-		if (!usb->buff[i]) {
-			pr_err("usb buffer kzalloc() failed\n");
-			usb_free_urb(usb->urb_r[i]);
-			usb->urb_r[i] = (struct urb *) 0;
-			return -ENOMEM;
-		}
+	/* allocate reusable rx urb */
+	usb->urb_r = usb_alloc_urb(0, GFP_ATOMIC);
+	if (!usb->urb_r) {
+		pr_err("usb_alloc_urb() failed\n");
+		return -ENOMEM;
+	}
+	usb->buff = kzalloc(USB_NET_BUFSIZ, GFP_ATOMIC);
+	if (!usb->buff) {
+		pr_err("usb buffer kzalloc() failed\n");
+		usb_free_urb(usb->urb_r);
+		usb->urb_r = (struct urb *) 0;
+		return -ENOMEM;
 	}
 
 	pr_debug("usb_net_raw_setup_ip_rx_urb }\n");
@@ -794,7 +796,6 @@ static int usb_net_raw_ip_setup_rx_urb( struct baseband_usb *usb)
 
 static void usb_net_raw_ip_free_rx_urb(struct baseband_usb *usb)
 {
-	int i;
 	pr_debug("usb_net_raw_ip_free_rx_urb {\n");
 
 	/* check input */
@@ -803,16 +804,14 @@ static void usb_net_raw_ip_free_rx_urb(struct baseband_usb *usb)
 		return;
 	}
 
-	for (i = 0; i < RAWIP_RX_BUFS; i++) {
-		/* free reusable rx urb */
-		if (usb->urb_r[i]) {
-			pr_debug("%s: freeing urb %p\n", __func__,
-							usb->urb_r[i]);
-			usb_free_urb(usb->urb_r[i]);
-		}
-
-		kfree(usb->buff[i]);
-		usb->buff[i] = NULL;
+	/* free reusable rx urb */
+	if (usb->urb_r) {
+		usb_free_urb(usb->urb_r);
+		usb->urb_r = (struct urb *) 0;
+	}
+	if (usb->buff) {
+		kfree(usb->buff);
+		usb->buff = (void *) 0;
 	}
 
 	pr_debug("usb_net_raw_ip_free_rx_urb }\n");
@@ -1027,7 +1026,7 @@ err_exit:
 
 static int usb_net_raw_ip_init(void)
 {
-	int i, j;
+	int i;
 	int err;
 	char name[32];
 
@@ -1121,14 +1120,13 @@ error_exit:
 				baseband_usb_net[i]->usb.tx_urb
 					= (struct urb *) 0;
 			}
-
 			/* stop usb rx */
-			for (j = 0; j < RAWIP_RX_BUFS; j++)
-				if (baseband_usb_net[i]->urb_r[j])
-					usb_kill_urb(baseband_usb_net[i]->
-								urb_r[j]);
+			if (baseband_usb_net[i]->usb.rx_urb) {
+				usb_kill_urb(baseband_usb_net[i]->usb.rx_urb);
+				baseband_usb_net[i]->usb.rx_urb
+					= (struct urb *) 0;
+			}
 			usb_net_raw_ip_free_rx_urb(baseband_usb_net[i]);
-
 			/* close usb */
 			baseband_usb_close(baseband_usb_net[i]);
 			baseband_usb_net[i] = (struct baseband_usb *) 0;
@@ -1141,59 +1139,42 @@ error_exit:
 
 static void usb_net_raw_ip_exit(void)
 {
-	struct baseband_usb *usb;
-	int i, j;
+	int i;
 
 	pr_debug("usb_net_raw_ip_exit {\n");
 
 	/* destroy multiple raw-ip network devices */
 	for (i = 0; i < max_intfs; i++) {
-		usb = baseband_usb_net[i];
-		/* stop baseband usb urbs */
-		if (usb) {
-			if (usb->usb.interface)
-				usb_autopm_get_interface(usb->usb.interface);
-
-			if (usb->usb.tx_workqueue)
-				cancel_work_sync(&usb->usb.tx_work);
-
-			/* stop usb tx */
-			pr_debug("%s: kill tx urb\n", __func__);
-			if (usb->usb.tx_urb) {
-				usb_kill_urb(usb->usb.tx_urb);
-				usb->usb.tx_urb = NULL;
-			}
-
-			/* stop usb rx */
-			pr_debug("%s: kill rx urbs\n", __func__);
-			for (j = 0; j < RAWIP_RX_BUFS; j++) {
-				if (usb->urb_r[j])
-					usb_kill_urb(usb->urb_r[j]);
-			}
-			usb_net_raw_ip_free_rx_urb(usb);
-
-			if (usb->usb.interface)
-				usb_autopm_put_interface(usb->usb.interface);
-		}
-
-		pr_debug("%s: unregister netdev\n", __func__);
 		/* unregister network device */
 		if (usb_net_raw_ip_dev[i]) {
 			unregister_netdev(usb_net_raw_ip_dev[i]);
 			free_netdev(usb_net_raw_ip_dev[i]);
 			usb_net_raw_ip_dev[i] = (struct net_device *) 0;
 		}
-
-		if (usb) {
-			pr_debug("%s: destroy tx workqueue\n", __func__);
-			if (usb->usb.tx_workqueue) {
-				destroy_workqueue(usb->usb.tx_workqueue);
-				usb->usb.tx_workqueue = NULL;
+		/* close baseband usb */
+		if (baseband_usb_net[i]) {
+			/* stop usb tx */
+			if (baseband_usb_net[i]->usb.tx_workqueue) {
+				destroy_workqueue(baseband_usb_net[i]
+					->usb.tx_workqueue);
+				baseband_usb_net[i]->usb.tx_workqueue
+					= (struct workqueue_struct *) 0;
 			}
-
+			if (baseband_usb_net[i]->usb.tx_urb) {
+				usb_kill_urb(baseband_usb_net[i]->usb.tx_urb);
+				baseband_usb_net[i]->usb.tx_urb
+					= (struct urb *) 0;
+			}
+			/* stop usb rx */
+			if (baseband_usb_net[i]->usb.rx_urb) {
+				usb_kill_urb(baseband_usb_net[i]->usb.rx_urb);
+				baseband_usb_net[i]->usb.rx_urb
+					= (struct urb *) 0;
+			}
+			usb_net_raw_ip_free_rx_urb(baseband_usb_net[i]);
 			/* close usb */
-			baseband_usb_close(usb);
-			baseband_usb_net[i] = NULL;
+			baseband_usb_close(baseband_usb_net[i]);
+			baseband_usb_net[i] = (struct baseband_usb *) 0;
 		}
 	}
